Reorganize the contents of $BUILD_TMP
In preparation for building the toolchain on the host, check for
existence of a stamp file before executing a build step and create
such file after the build step succeedes.

Define the subdirectories of $BUILD_TMP to be used by its sub-processes
in the main build-all.sh script.

Improve the diagnostic messages.