Pass the correct INSTALL_DIR location to `docker run`
Do not choose INSTALL_DIR automatically based on the availability of
the TOOLCHAIN_BUILD_IN_DOCKER environment variable. While setting that
marker variable in Dockerfile works transparently for most of the
scripts, it does not work for build.sh launcher.

This commit fixes `docker run` being invoked with

    --tmpfs "$HOST_BUILD_INSTALL_DIR:rw,exec,size=2G"

argument instead of the intended one:

    --tmpfs "$DOCKER_BUILD_INSTALL_DIR:rw,exec,size=2G"